+++ scripts/format.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 set -e
 
-black datamodel_code_generator tests
+black datamodel_code_generator tests --exclude tests/data
 isort --recursive datamodel_code_generator tests
