+++ tests/data/expected/main/main_jsonschema/output.py
@@ -1,0 +1,19 @@
+# generated by datamodel-codegen:
+#   filename:  person.json
+#   timestamp: 2019-07-26T00:00:00+00:00
+
+from __future__ import annotations
+
+from typing import Any, List, Optional
+
+from pydantic import BaseModel, Field, conint
+
+
+class Person(BaseModel):
+    firstName: Optional[str] = Field(None, description="The person's first name.")
+    lastName: Optional[str] = Field(None, description="The person's last name.")
+    age: Optional[conint(ge=0)] = Field(
+        None, description='Age in years which must be equal to or greater than zero.'
+    )
+    friends: Optional[List] = None
+    comment: Optional[Any] = None
