+++ tests/data/expected/main/main_jsonschema_nested_deep/__init__.py
@@ -1,0 +1,18 @@
+# generated by datamodel-codegen:
+#   filename:  nested_person.json
+#   timestamp: 2019-07-26T00:00:00+00:00
+
+from __future__ import annotations
+
+from typing import Optional
+
+from pydantic import BaseModel
+
+from .empty_parent.nested import deep as deep_1
+from .nested import deep
+
+
+class NestedPerson(BaseModel):
+    nested_deep_childJson: Optional[deep.Json] = None
+    nested_deep_childAnother: Optional[deep.Another] = None
+    empty_parent_nested_deep_childJson: Optional[deep_1.Json] = None
