+++ tests/data/expected/main/main_jsonschema_nested_deep/nested/deep.py
@@ -1,0 +1,17 @@
+# generated by datamodel-codegen:
+#   filename:  nested_person.json
+#   timestamp: 2019-07-26T00:00:00+00:00
+
+from __future__ import annotations
+
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class Json(BaseModel):
+    firstName: Optional[str] = None
+
+
+class Another(BaseModel):
+    firstName: Optional[str] = None
