+++ tests/data/expected/main/main_modular/foo/bar.py
@@ -1,0 +1,21 @@
+# generated by datamodel-codegen:
+#   filename:  modular.yaml
+#   timestamp: 1985-10-26T08:21:00+00:00
+
+from __future__ import annotations
+
+from typing import Any, Dict, List, Optional
+
+from pydantic import BaseModel
+
+
+class Thing(BaseModel):
+    attributes: Optional[Dict[str, Any]] = None
+
+
+class Thang(BaseModel):
+    attributes: Optional[List[Dict[str, Any]]] = None
+
+
+class Clone(Thing):
+    pass
