+++ tests/data/expected/main/main_modular/woo/__init__.py
@@ -1,0 +1,3 @@
+# generated by datamodel-codegen:
+#   filename:  modular.yaml
+#   timestamp: 1985-10-26T08:21:00+00:00
