+++ tests/data/expected/main/main_yaml/output.py
@@ -1,0 +1,16 @@
+# generated by datamodel-codegen:
+#   filename:  pet.yaml
+#   timestamp: 2019-07-26T00:00:00+00:00
+
+from __future__ import annotations
+
+from pydantic import BaseModel
+
+
+class Pet(BaseModel):
+    name: str
+    age: int
+
+
+class Model(BaseModel):
+    Pet: Pet
