+++ tests/test_main.py
@@ -7,7 +7,6 @@ import pytest
 from _pytest.capture import CaptureFixture
 from _pytest.tmpdir import TempdirFactory
 from freezegun import freeze_time
-from prance import ValidationError
 
 from datamodel_code_generator.__main__ import Exit, main
 
@@ -16,7 +15,7 @@ OPEN_API_DATA_PATH: Path = DATA_PATH / 'openapi'
 JSON_SCHEMA_DATA_PATH: Path = DATA_PATH / 'jsonschema'
 JSON_DATA_PATH: Path = DATA_PATH / 'json'
 YAML_DATA_PATH: Path = DATA_PATH / 'yaml'
-
+EXPECTED_MAIN_PATH = DATA_PATH / 'expected' / 'main'
 
 TIMESTAMP = '1985-10-26T01:21:00-07:00'
 
@@ -36,76 +35,7 @@ def test_main():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List[Pet]
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List[User]
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional[Event] = None
-'''
+            == (EXPECTED_MAIN_PATH / 'main' / 'output.py').read_text()
         )
 
     with pytest.raises(SystemExit):
@@ -129,78 +59,7 @@ def test_main_base_class():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, Field
-
-from custom_module import Base
-
-
-class Pet(Base):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(Base):
-    __root__: List[Pet]
-
-
-class User(Base):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(Base):
-    __root__: List[User]
-
-
-class Id(Base):
-    __root__: str
-
-
-class Rules(Base):
-    __root__: List[str]
-
-
-class Error(Base):
-    code: int
-    message: str
-
-
-class Api(Base):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(Base):
-    __root__: List[Api]
-
-
-class Event(Base):
-    name: Optional[str] = None
-
-
-class Result(Base):
-    event: Optional[Event] = None
-'''
+            == (EXPECTED_MAIN_PATH / 'main_base_class' / 'output.py').read_text()
         )
 
     with pytest.raises(SystemExit):
@@ -224,74 +83,7 @@ def test_target_python_version():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List['Pet']
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List['User']
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(BaseModel):
-    __root__: List['Api']
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional['Event'] = None
-'''
+            == (EXPECTED_MAIN_PATH / 'target_python_version' / 'output.py').read_text()
         )
 
     with pytest.raises(SystemExit):
@@ -315,26 +107,7 @@ def test_main_autodetect():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  person.json
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import Any, List, Optional
-
-from pydantic import BaseModel, Field, conint
-
-
-class Person(BaseModel):
-    firstName: Optional[str] = Field(None, description="The person\'s first name.")
-    lastName: Optional[str] = Field(None, description="The person\'s last name.")
-    age: Optional[conint(ge=0)] = Field(
-        None, description='Age in years which must be equal to or greater than zero.'
-    )
-    friends: Optional[List] = None
-    comment: Optional[Any] = None
-'''
+            == (EXPECTED_MAIN_PATH / 'main_autodetect' / 'output.py').read_text()
         )
 
     with pytest.raises(SystemExit):
@@ -382,26 +155,7 @@ def test_main_jsonschema():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  person.json
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import Any, List, Optional
-
-from pydantic import BaseModel, Field, conint
-
-
-class Person(BaseModel):
-    firstName: Optional[str] = Field(None, description="The person\'s first name.")
-    lastName: Optional[str] = Field(None, description="The person\'s last name.")
-    age: Optional[conint(ge=0)] = Field(
-        None, description='Age in years which must be equal to or greater than zero.'
-    )
-    friends: Optional[List] = None
-    comment: Optional[Any] = None
-'''
+            == (EXPECTED_MAIN_PATH / 'main_jsonschema' / 'output.py').read_text()
         )
     with pytest.raises(SystemExit):
         main()
@@ -430,67 +184,31 @@ def test_main_jsonschema_nested_deep():
             ]
         )
         assert return_code == Exit.OK
-        print(list(Path(output_dir).iterdir()))
         assert (
             output_init_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  nested_person.json
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-from .empty_parent.nested import deep as deep_1
-from .nested import deep
-
-
-class NestedPerson(BaseModel):
-    nested_deep_childJson: Optional[deep.Json] = None
-    nested_deep_childAnother: Optional[deep.Another] = None
-    empty_parent_nested_deep_childJson: Optional[deep_1.Json] = None
-'''
+            == (
+                EXPECTED_MAIN_PATH / 'main_jsonschema_nested_deep' / '__init__.py'
+            ).read_text()
         )
 
         assert (
             output_nested_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  nested_person.json
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-
-class Json(BaseModel):
-    firstName: Optional[str] = None
-
-
-class Another(BaseModel):
-    firstName: Optional[str] = None
-'''
+            == (
+                EXPECTED_MAIN_PATH
+                / 'main_jsonschema_nested_deep'
+                / 'nested'
+                / 'deep.py'
+            ).read_text()
         )
         assert (
             output_empty_parent_nested_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  nested_person.json
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-
-class Json(BaseModel):
-    firstName: Optional[str] = None
-'''
+            == (
+                EXPECTED_MAIN_PATH
+                / 'main_jsonschema_nested_deep'
+                / 'empty_parent'
+                / 'nested'
+                / 'deep.py'
+            ).read_text()
         )
     with pytest.raises(SystemExit):
         main()
@@ -513,23 +231,7 @@ def test_main_json():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  pet.json
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from pydantic import BaseModel
-
-
-class Pet(BaseModel):
-    name: str
-    age: int
-
-
-class Model(BaseModel):
-    Pet: Pet
-'''
+            == (EXPECTED_MAIN_PATH / 'main_json' / 'output.py').read_text()
         )
     with pytest.raises(SystemExit):
         main()
@@ -571,234 +273,13 @@ def test_main_yaml():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  pet.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from pydantic import BaseModel
-
-
-class Pet(BaseModel):
-    name: str
-    age: int
-
-
-class Model(BaseModel):
-    Pet: Pet
-'''
+            == (EXPECTED_MAIN_PATH / 'main_yaml' / 'output.py').read_text()
         )
     with pytest.raises(SystemExit):
         main()
 
 
-@pytest.mark.parametrize(
-    'expected',
-    [
-        {
-            (
-                '__init__.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-from . import models
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Result(BaseModel):
-    event: Optional[models.Event] = None
-
-
-class Source(BaseModel):
-    country: Optional[str] = None
-''',
-            (
-                'models.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from enum import Enum
-from typing import Any, Dict, List, Optional, Union
-
-from pydantic import BaseModel
-
-
-class Species(Enum):
-    dog = 'dog'
-    cat = 'cat'
-    snake = 'snake'
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-    species: Optional[Species] = None
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Event(BaseModel):
-    name: Optional[Union[str, float, int, bool, Dict[str, Any], List[str]]] = None
-''',
-            (
-                'collections.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-from . import models
-
-
-class Pets(BaseModel):
-    __root__: List[models.Pet]
-
-
-class Users(BaseModel):
-    __root__: List[models.User]
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-''',
-            (
-                'foo',
-                '__init__.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-from .. import Id
-
-
-class Tea(BaseModel):
-    flavour: Optional[str] = None
-    id: Optional[Id] = None
-
-
-class Cocoa(BaseModel):
-    quality: Optional[int] = None
-''',
-            (
-                'foo',
-                'bar.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import Any, Dict, List, Optional
-
-from pydantic import BaseModel
-
-
-class Thing(BaseModel):
-    attributes: Optional[Dict[str, Any]] = None
-
-
-class Thang(BaseModel):
-    attributes: Optional[List[Dict[str, Any]]] = None
-
-
-class Clone(Thing):
-    pass
-''',
-            (
-                'woo',
-                '__init__.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-''',
-            (
-                'woo',
-                'boo.py',
-            ): '''\
-# generated by datamodel-codegen:
-#   filename:  modular.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-from .. import Source, foo
-
-
-class Chocolate(BaseModel):
-    flavour: Optional[str] = None
-    source: Optional[Source] = None
-    cocoa: Optional[foo.Cocoa] = None
-''',
-        }
-    ],
-)
-def test_main_modular(
-    tmpdir_factory: TempdirFactory, expected: Mapping[str, str]
-) -> None:
+def test_main_modular(tmpdir_factory: TempdirFactory) -> None:
     """Test main function on modular file."""
 
     output_directory = Path(tmpdir_factory.mktemp('output'))
@@ -808,10 +289,10 @@ def test_main_modular(
 
     with freeze_time(TIMESTAMP):
         main(['--input', str(input_filename), '--output', str(output_path)])
-
-    for key, value in expected.items():
-        result = output_path.joinpath(*key).read_text()
-        assert result == value
+    main_modular_dir = EXPECTED_MAIN_PATH / 'main_modular'
+    for path in main_modular_dir.rglob('*.py'):
+        result = output_path.joinpath(path.relative_to(main_modular_dir)).read_text()
+        assert result == path.read_text()
 
 
 def test_main_modular_no_file() -> None:
@@ -836,83 +317,7 @@ def test_main_modular_filename(tmpdir_factory: TempdirFactory) -> None:
     )
 
 
-@pytest.mark.parametrize(
-    'expected',
-    [
-        '''\
-# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List[Pet]
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List[User]
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional[Event] = None
-'''
-    ],
-)
-def test_main_no_file(capsys: CaptureFixture, expected: str) -> None:
+def test_main_no_file(capsys: CaptureFixture) -> None:
     """Test main function on non-modular file with no output name."""
 
     input_filename = OPEN_API_DATA_PATH / 'api.yaml'
@@ -921,79 +326,13 @@ def test_main_no_file(capsys: CaptureFixture, expected: str) -> None:
         main(['--input', str(input_filename)])
 
     captured = capsys.readouterr()
-    assert captured.out == expected
+    assert (
+        captured.out == (EXPECTED_MAIN_PATH / 'main_no_file' / 'output.py').read_text()
+    )
     assert not captured.err
 
 
-@pytest.mark.parametrize(
-    'expected',
-    [
-        '''\
-# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 1985-10-26T08:21:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-
-class Pet(BaseModel):  # 1 2, 1 2, this is just a pet
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List[Pet]
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List[User]
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = None
-    apiVersionNumber: Optional[str] = None
-    apiUrl: Optional[AnyUrl] = None
-    apiDocumentationUrl: Optional[AnyUrl] = None
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional[Event] = None
-'''
-    ],
-)
-def test_main_custom_template_dir(capsys: CaptureFixture, expected: str) -> None:
+def test_main_custom_template_dir(capsys: CaptureFixture) -> None:
     """Test main function with custom template directory."""
 
     input_filename = OPEN_API_DATA_PATH / 'api.yaml'
@@ -1013,7 +352,10 @@ def test_main_custom_template_dir(capsys: CaptureFixture, expected: str) -> None
         )
 
     captured = capsys.readouterr()
-    assert captured.out == expected
+    assert (
+        captured.out
+        == (EXPECTED_MAIN_PATH / 'main_custom_template_dir' / 'output.py').read_text()
+    )
     assert not captured.err
 
 
@@ -1035,99 +377,7 @@ def test_pyproject():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import (
-    annotations,
-)
-
-from typing import (
-    List,
-    Optional,
-)
-
-from pydantic import (
-    AnyUrl,
-    BaseModel,
-    Field,
-)
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List[Pet]
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List[User]
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[
-        str
-    ] = Field(
-        None,
-        description="To be used as a dataset parameter value",
-    )
-    apiVersionNumber: Optional[
-        str
-    ] = Field(
-        None,
-        description="To be used as a version parameter value",
-    )
-    apiUrl: Optional[
-        AnyUrl
-    ] = Field(
-        None,
-        description="The URL describing the dataset\'s fields",
-    )
-    apiDocumentationUrl: Optional[
-        AnyUrl
-    ] = Field(
-        None,
-        description="A URL to the API console for each API",
-    )
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional[
-        Event
-    ] = None
-'''
+            == (EXPECTED_MAIN_PATH / 'pyproject' / 'output.py').read_text()
         )
 
     with pytest.raises(SystemExit):
@@ -1150,76 +400,7 @@ def test_validation():
         assert return_code == Exit.OK
         assert (
             output_file.read_text()
-            == '''# generated by datamodel-codegen:
-#   filename:  api.yaml
-#   timestamp: 2019-07-26T00:00:00+00:00
-
-from __future__ import annotations
-
-from typing import List, Optional
-
-from pydantic import AnyUrl, BaseModel, Field
-
-
-class Pet(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Pets(BaseModel):
-    __root__: List[Pet]
-
-
-class User(BaseModel):
-    id: int
-    name: str
-    tag: Optional[str] = None
-
-
-class Users(BaseModel):
-    __root__: List[User]
-
-
-class Id(BaseModel):
-    __root__: str
-
-
-class Rules(BaseModel):
-    __root__: List[str]
-
-
-class Error(BaseModel):
-    code: int
-    message: str
-
-
-class Api(BaseModel):
-    apiKey: Optional[str] = Field(
-        None, description='To be used as a dataset parameter value'
-    )
-    apiVersionNumber: Optional[str] = Field(
-        None, description='To be used as a version parameter value'
-    )
-    apiUrl: Optional[AnyUrl] = Field(
-        None, description="The URL describing the dataset\'s fields"
-    )
-    apiDocumentationUrl: Optional[AnyUrl] = Field(
-        None, description='A URL to the API console for each API'
-    )
-
-
-class Apis(BaseModel):
-    __root__: List[Api]
-
-
-class Event(BaseModel):
-    name: Optional[str] = None
-
-
-class Result(BaseModel):
-    event: Optional[Event] = None
-'''
+            == (EXPECTED_MAIN_PATH / 'validation' / 'output.py').read_text()
         )
 
     with pytest.raises(SystemExit):
